// citpwawcl/src/app/verifica/models/azione.model.ts
export interface Azione {
  cfUtenteAzione: string;
  denomUtenteAzione: string;
  descrizioneAzione: string;
  dtAzione: number;
  fkTipoAzione: number;
  idAccertamento: number;
  idAzione: number;
  idIspezione2018: number;
  idSanzione: number;
  idVerifica: number;
  nomeDoc: string;
  nomeDocOriginale: string;
  uidIndex: string;
}