// citpwawcl/src/app/verifica/models/distributore.model.ts
export class Distributore {
  idDatoDistri: number;
  fkStatoDistrib: number;
  statoDistribDesc: string;
  fkTipoContratto: number;
  desTipoContrattoDistrib: string;
  idImportDistrib: number;
  fkCategoriaUtil: string;
  desCategoriaUtil: string;
  fkCombustibile: number;
  desCombustibile: string;
  codiceAssenzaCatast: string;
  desAssenzaCatast: string;
  fkUnitaMisura: number;
  desUnitaMisura: string;
  flgPfPg: string;
  cognomeDenom: string;
  nome: string;
  cfPiva: string;
  annoRif: string;
  nrMesiFattur: number;
  dug: string;
  indirizzo: string;
  civico: string;
  cap: string;
  istatComune: string;
  podPdr: string;
  consumoAnno: string;
  consumoMensile: string;
  meseRiferimento: string;
  consumoGiornaliero: string;
  giornoRiferimento: number;
  volumetria: string;
  flgPfPgFatt: string;
  cognomeDenomFatt: string;
  nomeFatt: string;
  cfPivaFatt: string;
  dugFatt: string;
  indirizzoFatt: string;
  civicoFatt: string;
  capFatt: string;
  istatComuneFatt: string;
}